write-plus (2)/novel_generator/requirements.txt
python-dotenv
openai
python-docx
